Module hermes_governance_platform::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

DefaultForAuthorityAccount

DefaultForPalletStorageVersion

DefaultMinimumHermesAmountForCreatingPoll

DefaultMinimumHermesVotingAmount

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Functions

__type_value_for_default_for_authority_account

__type_value_for_default_for_pallet_storage_version

__type_value_for_default_minimum_hermes_amount_for_creating_poll

__type_value_for_default_minimum_hermes_voting_amount

Type Definitions

AccountIdOf

AuthorityAccount

Account which has permissions for changing Hermes minimum amount for voting and creating a poll

HermesPollData

HermesVotings

A vote of a particular user for a particular poll

MinimumHermesAmountForCreatingPoll

MinimumHermesVotingAmount

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

PalletStorageVersion

Pallet storage version

Timestamp